# services/web-bff/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: HooshiX Web BFF Public API
  version: 2.0.0
  description: Canonical browser-facing REST contract owned by Web BFF. All examples use synthetic .test identities and non-production placeholder credentials.
servers:
  - url: /
security:
  - BrowserSession: []
paths:
...
  /api/v1/identity/erasure:
    post:
      operationId: requestSelfErasure
      summary: Irreversibly erase the authenticated user's account data
      description: Requires recent authentication. When MFA is active, a current TOTP or recovery-code proof is mandatory. Acceptance immediately revokes every browser session while durable participants complete asynchronously.
      security: [{ BrowserSession: [] }]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SelfErasureRequest' }
            example: { confirmation: ERASE_MY_ACCOUNT, mfaProof: { type: TOTP, code: '123456' } }
      x-hooshix-consumer-example: { method: POST, path: /api/v1/identity/erasure, body: { confirmation: ERASE_MY_ACCOUNT, mfaProof: { type: TOTP, code: '123456' } } }
      responses:
        '202':
          description: Erasure was durably accepted and all browser sessions were revoked.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SelfErasureAcceptedResponse' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/InvalidSession' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/ErasureRejected' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /api/v1/auth/session/bootstrap:
    post:
      operationId: bootstrapBrowserSession
      summary: Create a pre-authentication browser session
      description: Creates an HttpOnly session cookie and returns an in-memory CSRF token. Origin and Fetch Metadata are required.
      security: []
      parameters: [{ $ref: '#/components/parameters/OptionalCsrfToken' }]
      x-hooshix-consumer-example: { method: POST, path: /api/v1/auth/session/bootstrap }
      responses:
        '201':
          description: Pre-authentication session created.
          content: { application/json: { schema: { $ref: '#/components/schemas/SessionResponse' } } }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/auth/session/csrf:
    post:
      operationId: rotateBrowserCsrf
      summary: Recover an in-memory CSRF token after browser document reload
      description: Rotates the server-side browser session and CSRF state under exact same-origin Origin and Fetch Metadata checks. Existing Identity refresh credentials and MFA challenges remain encrypted server-side and are never returned to browser JavaScript.
      security: []
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/OptionalCsrfToken' }]
      x-hooshix-consumer-example: { method: POST, path: /api/v1/auth/session/csrf }
      responses:
        '200':
          description: Rotated browser session and in-memory CSRF token.
          content: { application/json: { schema: { $ref: '#/components/schemas/SessionResponse' } } }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/auth/local:
    post:
      operationId: loginWithLocalPassword
      summary: Authenticate a pre-authentication browser session with a local password
      description: Browser JavaScript sends neither trusted client address nor refresh credential; the approved edge and BFF retain those authorities.
      security: [{ BrowserSession: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/CsrfToken' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/LocalLoginRequest' } } } }
      x-hooshix-consumer-example: { method: POST, path: /api/v1/auth/local, body: { channel: EMAIL, contact: person@example.test, password: CorrectHorseBatteryStaple!2026 } }
      responses:
        '200': { description: Authenticated session state., content: { application/json: { schema: { $ref: '#/components/schemas/SessionResponse' } } } }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/InvalidSession' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/RegistrationRejected' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /api/v1/auth/mfa/complete:
    post:
      operationId: completeBrowserMfaAuthentication
      summary: Complete a cookie-bound MFA login challenge
      description: The opaque Identity challenge remains encrypted in the server-side BFF session and is never returned to browser JavaScript.
      security: [{ BrowserSession: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/CsrfToken' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/MfaProofRequest' } } } }
      x-hooshix-consumer-example: { method: POST, path: /api/v1/auth/mfa/complete, body: { type: TOTP, code: '123456' } }
      responses:
        '200': { description: MFA-authenticated browser session state., content: { application/json: { schema: { $ref: '#/components/schemas/SessionResponse' } } } }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/InvalidSession' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /api/v1/auth/logout:
    post:
      operationId: logoutBrowserSession
      summary: Revoke the authenticated browser session
      security: [{ BrowserSession: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/CsrfToken' }]
      x-hooshix-consumer-example: { method: POST, path: /api/v1/auth/logout }
      responses:
        '204': { description: Session revoked. }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/InvalidSession' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /api/v1/auth/session:
    get:
      operationId: getBrowserSessionState
      summary: Read non-sensitive browser session state
      security: [{ BrowserSession: [] }]
      x-hooshix-consumer-example: { method: GET, path: /api/v1/auth/session }
      responses:
        '200': { description: Current session state., content: { application/json: { schema: { $ref: '#/components/schemas/SessionStateResponse' } } } }
        '401': { $ref: '#/components/responses/InvalidSession' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
...
components:
  securitySchemes:
    BrowserSession:
      type: apiKey
      in: cookie
      name: __Host-sajtech-session
      description: HttpOnly same-origin session. Its value is never exposed in this contract or browser storage.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Caller-generated UUIDv4 idempotency identity. Replays use the same value.
      schema:
        type: string
        format: uuid
        pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$'
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: Current session-bound synchronizer token returned by a reviewed BFF response; do not persist it.
      schema: { type: string, minLength: 32, maxLength: 512 }
    OptionalCsrfToken:
      name: X-CSRF-Token
      in: header
      required: false
      description: Required only when the anonymous-capable request carries an existing BFF session cookie; do not persist it.
      schema: { type: string, minLength: 32, maxLength: 512 }
...
  schemas:
...
    UuidV4:
      type: string
      format: uuid
      pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$'
      example: 11111111-1111-4111-8111-111111111111
    LocalLoginRequest:
      type: object
      additionalProperties: false
      required: [channel, contact, password]
      example: { channel: EMAIL, contact: person@example.test, password: CorrectHorseBatteryStaple!2026 }
      properties:
        channel: { $ref: '#/components/schemas/RegistrationChannel' }
        contact: { type: string, minLength: 1, maxLength: 254, example: person@example.test }
        password: { type: string, minLength: 1, maxLength: 4096, writeOnly: true, example: CorrectHorseBatteryStaple!2026 }
    MfaProofRequest:
      example: { type: TOTP, code: '123456' }
      discriminator: { propertyName: type }
      oneOf:
        - type: object
          additionalProperties: false
          required: [type, code]
          properties:
            type: { type: string, const: TOTP, example: TOTP }
            code: { type: string, pattern: '^[0-9]{6}$', writeOnly: true, example: '123456' }
        - type: object
          additionalProperties: false
          required: [type, code]
          properties:
            type: { type: string, const: RECOVERY_CODE, example: RECOVERY_CODE }
            code: { type: string, pattern: '^[A-Z2-7]{4}(?:-[A-Z2-7]{4}){3}$', writeOnly: true, example: AAAA-BBBB-CCCC-DDDD }
...
    RegistrationChannel:
      type: string
      enum: [EMAIL, PHONE]
      example: EMAIL
...
    SelfErasureRequest:
      type: object
      additionalProperties: false
      required: [confirmation]
      example: { confirmation: ERASE_MY_ACCOUNT, mfaProof: { type: TOTP, code: '123456' } }
      properties:
        confirmation:
          type: string
          const: ERASE_MY_ACCOUNT
        mfaProof:
          anyOf:
            - { $ref: '#/components/schemas/MfaProofRequest' }
            - { type: 'null' }
    SelfErasureAcceptedResponse:
      type: object
      additionalProperties: false
      required: [erasureRequestId, state, participantPolicyVersion]
      example: { erasureRequestId: 11111111-1111-4111-8111-111111111111, state: REQUESTED, participantPolicyVersion: '1' }
      properties:
        erasureRequestId: { $ref: '#/components/schemas/UuidV4' }
        state:
          type: string
          enum: [REQUESTED, IN_PROGRESS, COMPLETED, BLOCKED_BY_LEGAL_HOLD, FAILED_RETRYABLE]
        participantPolicyVersion:
          type: string
          pattern: '^[1-9][0-9]{0,8}$'
          minLength: 1
          maxLength: 64
...
    SessionResponse:
      type: object
      additionalProperties: false
      required: [csrfToken, mode]
      example: { csrfToken: synthetic-csrf-token-not-a-secret, mode: PREAUTH }
      properties:
        csrfToken: { type: string, minLength: 32, maxLength: 512, description: In-memory CSRF token; do not persist., example: synthetic-csrf-token-not-a-secret }
        mode: { type: string, enum: [PREAUTH, MFA_PREAUTH, AUTHENTICATED_ONBOARDING, TENANT_AUTHENTICATED], example: PREAUTH }
    SessionStateResponse:
      type: object
      additionalProperties: false
      required: [mode, authenticated, tenantSelected]
      example: { mode: TENANT_AUTHENTICATED, authenticated: true, tenantSelected: true }
      properties:
        mode: { type: string, enum: [PREAUTH, MFA_PREAUTH, AUTHENTICATED_ONBOARDING, TENANT_AUTHENTICATED], example: TENANT_AUTHENTICATED }
        authenticated: { type: boolean, example: true }
        tenantSelected: { type: boolean, example: true }
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code]
      example: { type: urn:hooshix:problem:invalid-request, title: Invalid request, status: 400, code: INVALID_REQUEST, instance: /api/v1/example }
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
          maxLength: 120
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
          pattern: '^[A-Z][A-Z0-9_]{0,63}$'
        instance:
          type: string
          maxLength: 1024
  responses:
    InvalidRequest:
      description: Request validation failed.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    InvalidSession:
      description: A supplied browser session is invalid.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Browser-origin or session-bound request protection rejected the request.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    RegistrationRejected:
      description: Registration precondition was not satisfied. Response is non-enumerating.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    ErasureRejected:
      description: Account erasure was not accepted because a lifecycle, recent-authentication, or MFA precondition was not satisfied.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    PayloadTooLarge:
      description: Request exceeded the public body bound.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    UnsupportedMediaType:
      description: Multipart public requests are not supported.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    RateLimited:
      description: The applicable semantic security quota was exceeded.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    DependencyUnavailable:
      description: The BFF runtime or an authoritative downstream dependency is unavailable.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    InternalError:
      description: Unexpected BFF failure with no implementation detail exposure.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }